umount-usb.sh: fixed messages in unmount_usb()

+++ umount-usb.sh
@@ -7,7 +7,7 @@
 # Author       : Copyright © 2025, Richard B. Romig, Mosfanet
 # Email        : rick.romig@gmail.com | rick.romig@mymetronet.com
 # Created      : 12 Mar 2025
-# Last updated : 05 Dec 2025
+# Last updated : 06 Dec 2025
 # Comments     : Only unmounts devices belonging to the current user.
 # TODO (Rick)  :
 # License      : GNU General Public License, version 2.0
@@ -45,15 +45,15 @@ unmount_usb() {
 	check_usb || { printf "No USB drive connected!\n" >&2; return; }
 	user=$(whoami)
 	usb_drive=$(find /media/"$user" -maxdepth 1 -type d -user "$user" | fzf --height 40% --reverse --prompt "Select the USB drive to unmount: ")
-	[[ "$usb_drive" ]] || { printf "No USB drives mounted!\n" >&2; return; }
+	[[ "$usb_drive" ]] || { printf "No USB drives mounted or selected!\n" >&2; return; }
 	umount -l "$usb_drive"
-	printf "\"%s\" unmounted or selected.\n" "$usb_drive"
+	printf "\"%s\" unmounted.\n" "$usb_drive"
 	[[ -d "$usb_drive" ]] && rmdir "$usb_drive"
 }
 
 main() {
 	local script="${0##*/}"
-	local version="2.0.25339"
+	local version="2.0.25340"
 	check_package fzf
 	unmount_usb
 	over_line "$script $version"
